requirements.txt
aioredis==1.2.0
amqp==2.3.2
argon2-cffi==18.3.0
asgiref==2.3.2
asn1crypto==0.24.0
astroid==2.1.0
astunparse==1.6.1
async-timeout==3.0.1
atomicwrites==1.2.1
attrs==18.2.0
autobahn==18.12.1
Automat==0.7.0
backcall==0.1.0
billiard==3.5.0.5
bleach==3.0.2
CacheControl==0.12.5
cachetools==3.0.0
cairocffi==0.9.0
CairoSVG==2.2.1
calysto==1.0.6
celery==4.2.1
certifi==2018.11.29
cffi==1.11.5
channels==2.1.6
channels-redis==2.3.2
chardet==3.0.4
click==7.0
colour==0.1.5
constantly==15.1.0
cryptography==2.4.2
cssselect2==0.2.1
daphne==2.2.4
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.5
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-debug-toolbar==1.11
django-filter==2.0.0
django-js-asset==1.1.0
django-render-block==0.5
django-taggit==0.23.0
django-templated-email==2.2.0
djangorestframework==3.9.0
entrypoints==0.2.3
Faker==1.0.1
firebase-admin==2.14.0
Flask==1.0.2
Flask-Cors==3.0.7
future==0.17.1
geojson==2.4.1
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.13.2
google-resumable-media==0.3.2
googleapis-common-protos==1.5.5
hiredis==0.3.1
html5lib==1.0.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
jupyter-http-over-ws==0.0.2
kombu==4.2.2.post1
lazy-object-proxy==1.3.1
lxml==4.3.0
Markdown==3.0.1
MarkupSafe==1.1.0
matlab-kernel==0.16.1
mccabe==0.6.1
metakernel==0.20.14
mistune==0.8.4
module-name==0.5.1
more-itertools==5.0.0
mpld3==0.3
msgpack==0.6.0
msgpack-python==0.5.6
nbconvert==5.4.0
nbformat==4.4.0
ndg-httpsclient==0.5.1
notebook==5.7.4
numpy==1.15.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pixiedust==1.1.15
pluggy==0.8.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.6.1
psutil==5.4.8
psycopg2==2.7.6.1
ptyprocess==0.6.0
pusher==2.1.2
py==1.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
pyglet==1.3.2
Pygments==2.3.1
PyHamcrest==1.9.0
pylast==3.0.0
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
pyOpenSSL==18.0.0
pyparsing==2.3.0
pytest==4.1.0
pytest-asyncio==0.9.0
pytest-django==3.4.4
python-dateutil==2.7.5
python3-xlib==0.15
pytz==2018.9
pyzmq==17.1.2
redis==3.0.1
requests==2.21.0
rsa==4.0
scikit-learn==0.20.2
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
six==1.12.0
sklearn==0.0
soupy==0.3
sqlparse==0.2.4
stripe==2.17.0
svgwrite==1.2.1
terminado==0.8.1
testpath==0.4.2
text-unidecode==1.2
tflearn==0.3.2
tinycss2==0.6.1
tornado==5.1.1   
traitlets==4.3.2
Twisted==18.9.0
txaio==18.8.1
typed-ast==1.1.1
urllib3==1.24.1
vine==1.2.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
zope.interface==4.6.0
